#30 모달 컴포넌트화
