+++ src/pages/Inventory/index.tsx
@@ -4,6 +4,7 @@ import { useRecoilState } from "recoil";
 import { InventoryDto } from "../../recoil/DBAtom";
 import { IngredientDto } from "../../interface/DataInterface";
 import { useState } from "react";
+import { addIngredient } from "./addIngredient";
 
 export const Inventory = () => {
   // 요청을 보낼 URL
@@ -11,7 +12,7 @@ export const Inventory = () => {
 
   // 토큰 값
   const token =
-    "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJzdHJpbmciLCJpZCI6IjIiLCJleHAiOjE2OTYxNzg4MTIsInVzZXJuYW1lIjoic3RyaW5nIn0.OJs9xlIbgKoa0T7Rs09vbOS3NrRW_ejUcyzkXqHyY264OUCOIZ8j_nFlqqa4FX87J3rmfPpon8JLzA7gWj9jzg";
+    "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJzdHJpbmciLCJpZCI6IjIiLCJleHAiOjE2OTYyMTgxNzMsInVzZXJuYW1lIjoic3RyaW5nIn0.e25j6RwehGzfwnQdHy2H7d_5WayW8nlkTF2DoXW4MLkX_0U4FHINQOwibYVVeBYjliqRIeIm5vbuj4pbWQKLQA";
   const [Inv, setInv] = useRecoilState(InventoryDto);
   // Axios 인스턴스 생성
   const instance: AxiosInstance = axios.create({
@@ -48,196 +49,20 @@ export const Inventory = () => {
   };
 
   const [add, setAdd] = useState(false);
-  const addIngredient = () => {
-    setAdd(!add);
-    console.log(add);
+  const isaddIngredient = () => {
+    setAdd(true);
   };
-  const modalblur = () => {
-    return add
-      ? "fixed inset-0 bg-gray-500 bg-opacity-50 transition-opacity"
-      : "";
-  };
-  const inputcss = () => {
-    return " px-4 py-2 border border-gray-300 rounded focus:outline-none focus:border-red-300 focus:border-2 ";
-  };
-  const today = new Date().toISOString().split("T")[0];
 
   return (
     <div className="flex justify-center flex-col mt-28 items-left ml-[10rem]">
       <div className="w-[100rem] overflow-x-auto sm:-mx-6 lg:-mx-8 border-4 rounded-md px-4 py-4">
-        {add && (
-          <div className={modalblur()}>
-            <div className="flex justify-center flex-col mt-[20rem] items-left ml-[10rem]">
-              <div className="w-[100rem] overflow-x-auto sm:-mx-6 lg:-mx-8 border-4 rounded-md px-4 py-4 bg-white">
-                <table className="min-w-full text-center text-sm font-light border-t-2 border-black bg-white">
-                  <thead className="border-b bg-neutral-50 font-medium dark:border-neutral-500 dark:text-neutral-800 ">
-                    <tr>
-                      <th
-                        scope="col"
-                        className=" px-6 py-4 text-right mr-4"
-                      ></th>
-                      <th
-                        scope="col"
-                        className=" px-6 py-4 text-right mr-4"
-                      ></th>
-                      <th
-                        scope="col"
-                        className="whitespace-nowrap px-6 py-4 text-right"
-                      >
-                        재료명
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right">
-                        남은 재고
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right">
-                        단위
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right">
-                        구매 날짜
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        구매 예정 날짜
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        재고 경고량
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        구매 가격(원가)
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        구매 묶음 양
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        소비 기한
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        이상적인 양
-                      </th>
-                      <th scope="col" className=" px-6 py-4 text-right mr-4">
-                        구매 링크
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr className="border-b dark:border-neutral-500">
-                      <td>
-                        <button
-                          className={eraseButtonDesign()}
-                          onClick={() => {
-                            setAdd(false);
-                          }}
-                        >
-                          취소
-                        </button>
-                      </td>
-                      <td>
-                        <button className={buttonDesign()}>저장</button>
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 font-medium text-right">
-                        <input
-                          type="text"
-                          onChange={() => {
-                            console.log("text");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("remainQuantity");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <select className={inputcss()}>
-                          <option value="g">g</option>
-                          <option value="ea">ea</option>
-                        </select>
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="date"
-                          className={inputcss()}
-                          defaultValue={today}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="date"
-                          className={inputcss()}
-                          defaultValue={today}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("alertquantity");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("alertquantity");
-                          }}
-                          className={inputcss()}
-                        />
-                        원
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("buyquantity");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("alertquantity");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="number"
-                          onChange={() => {
-                            console.log("alertquantity");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                      <td className="whitespace-nowrap  px-6 py-4 text-right">
-                        <input
-                          type="text"
-                          onChange={() => {
-                            console.log("text");
-                          }}
-                          className={inputcss()}
-                        />
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
-        )}
+        {/* 재료 추가버튼 클릭시 뜨는 모달 컴포넌트 */}
+        {add && addIngredient(add, setAdd)}
         <div className="flex items-center ml-[1rem] mb-[1rem]">
           <button
             id="재료추가"
             className={buttonDesign()}
-            onClick={addIngredient}
+            onClick={isaddIngredient}
           >
             재료추가
           </button>
